move attribute mapping to a base class
